backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-Limiter==3.8.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
marshmallow==3.21.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
pytest==8.2.2
pytest-flask==1.3.0
